mammography/requirements/main.txt
# SHA1:6860f81429b7f52990f11ed51f5e922d192cf4bf
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
aiohttp==3.8.3
    # via
    #   -c mammography/requirements/dev.txt
    #   fsspec
aiosignal==1.3.1
    # via
    #   -c mammography/requirements/dev.txt
    #   aiohttp
antlr4-python3-runtime==4.9.3
    # via
    #   -c mammography/requirements/dev.txt
    #   hydra-core
    #   omegaconf
async-timeout==4.0.2
    # via
    #   -c mammography/requirements/dev.txt
    #   aiohttp
attrs==22.2.0
    # via
    #   -c mammography/requirements/dev.txt
    #   aiohttp
certifi==2022.12.7
    # via
    #   -c mammography/requirements/dev.txt
    #   requests
    #   sentry-sdk
charset-normalizer==2.1.1
    # via
    #   -c mammography/requirements/dev.txt
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   -c mammography/requirements/dev.txt
    #   wandb
docker-pycreds==0.4.0
    # via
    #   -c mammography/requirements/dev.txt
    #   wandb
frozenlist==1.3.3
    # via
    #   -c mammography/requirements/dev.txt
    #   aiohttp
    #   aiosignal
fsspec[http]==2022.11.0
    # via
    #   -c mammography/requirements/dev.txt
    #   lightning-lite
    #   pytorch-lightning
gitdb==4.0.10
    # via
    #   -c mammography/requirements/dev.txt
    #   gitpython
gitpython==3.1.30
    # via
    #   -c mammography/requirements/dev.txt
    #   wandb
hydra-core==1.3.1
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
idna==3.4
    # via
    #   -c mammography/requirements/dev.txt
    #   requests
    #   yarl
lightning-lite==1.8.6
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
lightning-utilities==0.5.0
    # via
    #   -c mammography/requirements/dev.txt
    #   lightning-lite
    #   pytorch-lightning
multidict==6.0.4
    # via
    #   -c mammography/requirements/dev.txt
    #   aiohttp
    #   yarl
numpy==1.24.1
    # via
    #   -c mammography/requirements/dev.txt
    #   lightning-lite
    #   opencv-python-headless
    #   pandas
    #   pylibjpeg
    #   pylibjpeg-libjpeg
    #   pytorch-lightning
    #   tensorboardx
    #   torchmetrics
    #   torchvision
omegaconf==2.3.0
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
    #   hydra-core
opencv-python-headless==4.7.0.68
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
packaging==23.0
    # via
    #   -c mammography/requirements/dev.txt
    #   hydra-core
    #   lightning-lite
    #   lightning-utilities
    #   pytorch-lightning
    #   torchmetrics
pandas==1.5.2
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
pathtools==0.1.2
    # via
    #   -c mammography/requirements/dev.txt
    #   wandb
pillow==9.4.0
    # via
    #   -c mammography/requirements/dev.txt
    #   torchvision
promise==2.3
    # via
    #   -c mammography/requirements/dev.txt
    #   wandb
protobuf==3.20.1
    # via
    #   -c mammography/requirements/dev.txt
    #   tensorboardx
    #   wandb
psutil==5.9.4
    # via
    #   -c mammography/requirements/dev.txt
    #   wandb
pydicom==2.3.1
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
pylibjpeg==1.4.0
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
pylibjpeg-libjpeg==1.3.2
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
python-dateutil==2.8.2
    # via
    #   -c mammography/requirements/dev.txt
    #   pandas
pytorch-lightning==1.8.6
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
pytz==2022.7.1
    # via
    #   -c mammography/requirements/dev.txt
    #   pandas
pyyaml==6.0
    # via
    #   -c mammography/requirements/dev.txt
    #   omegaconf
    #   pytorch-lightning
    #   wandb
requests==2.28.2
    # via
    #   -c mammography/requirements/dev.txt
    #   fsspec
    #   torchvision
    #   wandb
sentry-sdk==1.13.0
    # via
    #   -c mammography/requirements/dev.txt
    #   wandb
setproctitle==1.3.2
    # via
    #   -c mammography/requirements/dev.txt
    #   wandb
shortuuid==1.0.11
    # via
    #   -c mammography/requirements/dev.txt
    #   wandb
six==1.16.0
    # via
    #   -c mammography/requirements/dev.txt
    #   docker-pycreds
    #   promise
    #   python-dateutil
smmap==5.0.0
    # via
    #   -c mammography/requirements/dev.txt
    #   gitdb
tensorboardx==2.5.1
    # via
    #   -c mammography/requirements/dev.txt
    #   pytorch-lightning
torch==1.11.0
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
    #   lightning-lite
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
torchmetrics==0.11.0
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
    #   pytorch-lightning
torchvision==0.12.0
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
tqdm==4.64.1
    # via
    #   -c mammography/requirements/dev.txt
    #   pytorch-lightning
typing-extensions==4.4.0
    # via
    #   -c mammography/requirements/dev.txt
    #   lightning-lite
    #   lightning-utilities
    #   pytorch-lightning
    #   torch
    #   torchvision
urllib3==1.26.14
    # via
    #   -c mammography/requirements/dev.txt
    #   requests
    #   sentry-sdk
wandb==0.13.7
    # via
    #   -c mammography/requirements/dev.txt
    #   -r mammography/requirements/main.in
yarl==1.8.2
    # via
    #   -c mammography/requirements/dev.txt
    #   aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
